src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appnope==0.1.0            # via ipython
attrs==17.4.0             # via pytest
backcall==0.1.0           # via ipython
click==6.7                # via pip-tools
coverage==4.5.1           # via pytest-cov
decorator==4.3.0          # via ipython, traitlets
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==2.0.7
django==2.0.4
djangorestframework==3.8.2
first==2.0.1              # via pip-tools
invoke==0.22.1
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==6.3.1            # via ipdb
jedi==0.12.0              # via ipython
mccabe==0.6.1             # via pylama
more-itertools==4.1.0     # via pytest
parso==0.2.0              # via jedi
pexpect==4.5.0            # via ipython
pickleshare==0.7.4        # via ipython
pillow==5.1.0
pip-tools==2.0.1
pluggy==0.6.0             # via pytest
prompt-toolkit==1.0.15    # via ipython
psycopg2==2.7.4
ptyprocess==0.5.2         # via pexpect
py==1.5.3                 # via pytest
pycodestyle==2.4.0        # via pylama
pydocstyle==2.1.1         # via pylama
pyflakes==1.6.0           # via pylama
pygments==2.2.0           # via ipython
pylama==7.4.3
pytest-cov==2.5.1
pytest-django==3.2.1
pytest-mock==1.9.0
pytest==3.5.1
pytz==2018.4
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via django-environ, django-extensions, more-itertools, pip-tools, prompt-toolkit, pydocstyle, pytest, traitlets
snowballstemmer==1.2.1    # via pydocstyle
sqlparse==0.2.4           # via django-debug-toolbar
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.14.1
